requirements.txt

fastapi
uvicorn

langchain
crewai       
sentence-transformers
faiss-cpu    
transformers
torch
huggingface-hub
numpy
pandas
PyMuPDF       
tqdm
requests

loguru

duckduckgo-search

ddgs
